add delete files from opensearch

+++ server/src/documents/documents.service.ts
@@ -4,12 +4,14 @@ import { CreateDocumentDto } from './dto/document.dto';
 import { Document } from './document.type';
 import { S3Service } from 'src/s3/s3.service';
 import { config } from 'src/config';
+import { SearchService } from 'src/search/search.service';
 
 @Injectable()
 export class DocumentsService {
   constructor(
     private prisma: PrismaService,
     private readonly s3: S3Service,
+    private readonly searchService: SearchService,
   ) {}
 
   async getUrl({
@@ -19,7 +21,8 @@ export class DocumentsService {
     fileName: string;
     fileType: string;
   }): Promise<{ presignedUrl: string; key: string }> {
-    const key = `uploads/${fileName}`;
+    const timestamp = Date.now();
+    const key = `uploads/${timestamp}_${fileName}`;
     const url = await this.s3.uploadFile(config.S3_BUCKET!, key, fileType);
     return { presignedUrl: url, key };
   }
@@ -44,8 +47,10 @@ export class DocumentsService {
         throw new Error('Document not found');
       }
 
-      const key = `uploads/${document.filename}`;
-      const res = await this.s3.deleteFile(config.S3_BUCKET!, key);
+      const res = await this.s3.deleteFile(
+        config.S3_BUCKET!,
+        document.objectKey!,
+      );
 
       const isSuccess =
         res.DeleteMarker === true ||
@@ -55,6 +60,16 @@ export class DocumentsService {
       if (!isSuccess) {
         throw new Error('Failed to delete file from S3');
       }
+
+      const opensearchId = await this.searchService.getDocumentId(
+        'documents',
+        document.objectKey!,
+      );
+      if (!opensearchId) {
+        throw new Error('No such document');
+      }
+
+      await this.searchService.deleteDocument('documents', opensearchId);
 
       return await this.prisma.document.delete({ where: { id } });
     } catch (e) {
